Fixes issue with category and guides
Guides were not displaying because of a missing ‘,’ in the meta tag
information in the guides view. Added a helper to generate the meta_tag
options

Also fixed and issue where the categories were not linking to the
correct article urls.

Gave the category factory a more proper title using Faker.